Moved animations into separate folder
