+++ src/edu/uco/sdd/rocketdog/model/AidItem.java
@@ -16,6 +16,7 @@
  */
 package edu.uco.sdd.rocketdog.model;
 
+import edu.uco.sdd.rocketdog.model.Animations.IAnimateStrategy;
 import javafx.geometry.Point2D;
 import javafx.geometry.Rectangle2D;
 import javafx.scene.image.Image;
