+++ src/edu/uco/sdd/rocketdog/model/Animations/IAnimateStrategy.java
@@ -1,9 +1,9 @@
-package edu.uco.sdd.rocketdog.model;
+package edu.uco.sdd.rocketdog.model.Animations;
 
 import javafx.geometry.Rectangle2D;
 import javafx.scene.image.Image;
 
-interface IAnimateStrategy {
+public interface IAnimateStrategy {
     void handle();
     Rectangle2D getCurrentView();
     Image getImage();
