+++ src/edu/uco/sdd/rocketdog/model/Animations/SpitzChargeGunAnimateStrategy.java
@@ -1,4 +1,4 @@
-package edu.uco.sdd.rocketdog.model;
+package edu.uco.sdd.rocketdog.model.Animations;
 
 import javafx.geometry.Rectangle2D;
