+++ src/edu/uco/sdd/rocketdog/model/BoostItem.java
@@ -5,6 +5,7 @@
  */
 package edu.uco.sdd.rocketdog.model;
 
+import edu.uco.sdd.rocketdog.model.Animations.BoostAnimateStrategy;
 import javafx.geometry.Point2D;
 import javafx.scene.image.ImageView;
 /**
